requirements.txt
quickfix==1.15.1
requests==2.31.0
python-dotenv==0.21.1
psycopg2==2.9.10
uuid==1.30
